+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Add base error class (`Chroma::Error`).
 * Change namespace for errors from `Chroma::Errors::` to `Chroma::`.
 * Add red, green and blue color attributes
+* Add hue, saturation, lightness color attributes
 
 ## v0.1.0 - 2016-05-26